重构数据源创建
