+++ lodsve-mybatis/src/main/java/lodsve/mybatis/datasource/DataSourceBeanFactoryPostProcessor.java
@@ -1,0 +1,119 @@
+/*
+ * Copyright (C) 2018  Sun.Hao
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package lodsve.mybatis.datasource;
+
+import com.p6spy.engine.spy.P6DataSource;
+import lodsve.core.properties.Profiles;
+import lodsve.mybatis.datasource.builder.RdbmsDataSourceBeanDefinitionBuilder;
+import lodsve.mybatis.datasource.dynamic.DynamicDataSource;
+import lodsve.mybatis.p6spy.LodsveP6OptionsSource;
+import lodsve.mybatis.properties.MyBatisProperties;
+import lodsve.mybatis.utils.Constants;
+import org.springframework.beans.BeansException;
+import org.springframework.beans.factory.config.BeanDefinition;
+import org.springframework.beans.factory.config.BeanFactoryPostProcessor;
+import org.springframework.beans.factory.config.ConfigurableListableBeanFactory;
+import org.springframework.beans.factory.support.BeanDefinitionBuilder;
+import org.springframework.beans.factory.support.DefaultListableBeanFactory;
+
+import java.util.*;
+
+/**
+ * 注册数据源.
+ *
+ * @author sunhao(sunhao.java @ gmail.com)
+ * @date 2018-2-8-0008 16:44
+ */
+public class DataSourceBeanFactoryPostProcessor implements BeanFactoryPostProcessor {
+    private MyBatisProperties myBatisProperties;
+
+    public DataSourceBeanFactoryPostProcessor(MyBatisProperties myBatisProperties) {
+        this.myBatisProperties = myBatisProperties;
+    }
+
+    @Override
+    public void postProcessBeanFactory(ConfigurableListableBeanFactory beanFactory) throws BeansException {
+        if (!(beanFactory instanceof DefaultListableBeanFactory)) {
+            return;
+        }
+
+        String[] dataSources = myBatisProperties.getDataSources();
+        Map<String, BeanDefinition> beanDefinitions = new HashMap<>(dataSources.length);
+
+        // 组装一些信息
+        DataSourceBean dataSourceBean = new DataSourceBean(dataSources);
+        String defaultDataSourceKey = dataSourceBean.getDefaultDataSourceKey();
+        beanDefinitions.putAll(dataSourceBean.getBeanDefinitions());
+
+        // 动态数据源
+        BeanDefinitionBuilder dynamicDataSource = BeanDefinitionBuilder.genericBeanDefinition(DynamicDataSource.class);
+        dynamicDataSource.addConstructorArgValue(dataSourceBean.getDataSourceNames());
+        dynamicDataSource.addConstructorArgValue(defaultDataSourceKey);
+
+        boolean p6spy = Profiles.getProfile("p6spy");
+        if (p6spy) {
+            // 使用p6spy
+            beanDefinitions.put(Constants.REAL_DATA_SOURCE_BEAN_NAME, dynamicDataSource.getBeanDefinition());
+
+            BeanDefinitionBuilder p6spyDataSource = BeanDefinitionBuilder.genericBeanDefinition(P6DataSource.class);
+            p6spyDataSource.addConstructorArgReference(Constants.REAL_DATA_SOURCE_BEAN_NAME);
+
+            beanDefinitions.put(Constants.DATA_SOURCE_BEAN_NAME, p6spyDataSource.getBeanDefinition());
+
+            // 初始化配置
+            LodsveP6OptionsSource.init();
+        } else {
+            beanDefinitions.put(Constants.DATA_SOURCE_BEAN_NAME, dynamicDataSource.getBeanDefinition());
+        }
+    }
+
+    private static class DataSourceBean {
+        private String defaultDataSourceKey;
+        private Map<String, BeanDefinition> beanDefinitions;
+        private List<String> dataSourceNames;
+
+        DataSourceBean(String[] dataSources) {
+            beanDefinitions = new HashMap<>(dataSources.length);
+            dataSourceNames = new ArrayList<>(dataSources.length);
+
+            for (int i = 0; i < dataSources.length; i++) {
+                String name = dataSources[i];
+                BeanDefinition dsBeanDefinition = new RdbmsDataSourceBeanDefinitionBuilder(name).build();
+
+                if (i == 0) {
+                    defaultDataSourceKey = name;
+                }
+
+                beanDefinitions.put(name, dsBeanDefinition);
+                dataSourceNames.add(name);
+            }
+        }
+
+        String getDefaultDataSourceKey() {
+            return defaultDataSourceKey;
+        }
+
+        Map<String, BeanDefinition> getBeanDefinitions() {
+            return beanDefinitions;
+        }
+
+        List<String> getDataSourceNames() {
+            return dataSourceNames;
+        }
+    }
+}
